+++ src/src/uni_electrolyte/evaluator/dataset/__init__.py
@@ -1,2 +1,4 @@
-from .pyG_thuEMol import pyG_thuEMol
+from .pyG_thuEMol import thuEMol
 from .spk_thuEMol import spk_thuEMol
+from .data_transform import *
+