+++ src/src/uni_electrolyte/evaluator/trainer/__init__.py
@@ -1,2 +1,2 @@
 from .pyG_trainer import pyG_trainer
-from .spk_task import ModelOutput, SinglePropertyClrTask, spk_inference_without_label
+from .spk_task import ModelOutput, SinglePropertyClrTask