+++ src/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [1.4.1](https://github.com/Franklalalala/uni_electrolyte_beta/compare/v1.4.0...v1.4.1) (2023-09-25)
+
+
+### Bug Fixes
+
+* change entry name ([bcff69f](https://github.com/Franklalalala/uni_electrolyte_beta/commit/bcff69ff9fc8d7ded69ea90961ad4c34a50b9cd4))
+
+## [1.4.0](https://github.com/Franklalalala/uni_electrolyte_beta/compare/v1.3.0...v1.4.0) (2023-09-25)
+
+
+### Features
+
+* add model entry ([a62ff61](https://github.com/Franklalalala/uni_electrolyte_beta/commit/a62ff61c65135aae71174c2b64240a42b964039b))
+
+## [1.3.0](https://github.com/Franklalalala/uni_electrolyte_beta/compare/v1.2.0...v1.3.0) (2023-09-24)
+
+
+### Features
+
+* clean property dependency ([cb80d93](https://github.com/Franklalalala/uni_electrolyte_beta/commit/cb80d9347096fded26ade71d5bd4f8181e15ba52))
+
+## [1.2.0](https://github.com/Franklalalala/uni_electrolyte_beta/compare/v1.1.1...v1.2.0) (2023-09-22)
+
+
+### Features
+
+* change predict step ([e7f4f4b](https://github.com/Franklalalala/uni_electrolyte_beta/commit/e7f4f4bd6c932100679f80d05fcd96d56b04f300))
+
 ## [1.1.1](https://github.com/Franklalalala/uni_electrolyte_beta/compare/v1.1.0...v1.1.1) (2023-09-21)